refactor(deleteallusers): server-sided only admins may delete other admins

+++ packages/server/src/controllers/adminsAndManagers/deleteAllUsers/index.ts
@@ -8,12 +8,13 @@ import isCountOneInUsers from '../../../utils/isCountOneInUsers';
 const deleteAllUsers: Route = async (req, res) => {
   try {
     const user = req?.user as UserObject;
+    const currentUserIsAdmin = user?.role === `admin`;
     const { deleteAdmins } = req?.body;
 
     await deleteAllUsersSchema.validate({ deleteAdmins });
     const isOnlyAdmin = await isCountOneInUsers(`role`, `admin`);
 
-    if (deleteAdmins && !isOnlyAdmin) {
+    if (currentUserIsAdmin && deleteAdmins && !isOnlyAdmin) {
       await User.destroy({
         where: {
           id: {
